% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\alias{ko15.CDF}
\title{MS data in CDF format}
\description{
This data set represents a single CDF file in (AIA/ANDI) NetCDF
format from a larger experiment in which the metabolic consequences
of knocking the fatty acid amide hydrolase (FAAH) gene in mice was
investigated. The file contains datain centroid mode acquired in
positive ion mode from 200-600 m/z and 2500-4500 seconds.

Data file:
\itemize{
\item \emph{ko15.CDF} file in NetCDF format.
}
}
\references{
\itemize{
\item Saghatelian, A et al. \emph{Assignment of endogenous substrates to enzymes
by global metabolite profiling}, Biochemistry, 2004.
http://dx.doi.org/10.1021/bi0480335
}
}
