% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faahKO.R
\name{faahKO}
\alias{faahKO}
\alias{ko15.CDF}
\alias{ko16.CDF}
\alias{ko18.CDF}
\alias{ko19.CDF}
\alias{ko21.CDF}
\alias{ko22.CDF}
\alias{wt15.CDF}
\alias{wt16.CDF}
\alias{wt18.CDF}
\alias{wt19.CDF}
\alias{wt21.CDF}
\alias{wt22.CDF}
\title{Saghatelian et al. (2004) FAAH knockout LC/MS data}
\usage{
faahKO()
}
\description{
Data from the article Assignment of endogenous substrates to enzymes by
global metabolite profiling Biochemistry; 2004; 43(45) by Saghatelian et al..
LC-MS data from 6 FAAHKO knock-out and 6 wild type mice acquired in
positive ionization mode. Data is subset to an m/z range of 200-600 and
retention time range from 2500-4500 seconds. Data files are provided in
NetCDF format. Text files provide information on individual samples (\emph{s_}),
the experimental condition (\emph{i_}).

Data available on Zenodo at
\href{https://doi.org/10.5281/zenodo.19606563}{doi: 10.5281/zenodo.19606563}

Data files can be loaded individually through their file name.
Use \code{faahKO()} to retrieve all data files and return their path.
}
\references{
Saghatelian A, Trauger SA, Want EJ, Hawkins EG, Siuzdak G and Cravatt BF.
Assignment of Endogenous Substrates to Enzymes by Global Metabolite
Profiling. Biochemistry 2004; 43 (45):14332-14339. doi: 10.1021/bi0480335
}
\author{
Johannes Rainer and Steffen Neumann
}
