% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTICR.R
\name{FTICR}
\alias{FTICR}
\alias{HAM004_641fE_14-11-07--Exp1.extracted.mzML}
\alias{HAM004_641fE_14-11-07--Exp2.extracted.mzML}
\alias{HAM004_641fE_14-11-07--Exp3.extracted.mzML}
\alias{HAM004_641fE_14-11-07--Exp4.extracted.mzML}
\alias{HAM004_641fE_14-11-07--Exp5.extracted.mzML}
\alias{HAM005_641fE_14-11-07--Exp1.extracted.mzML}
\alias{HAM005_641fE_14-11-07--Exp2.extracted.mzML}
\alias{HAM005_641fE_14-11-07--Exp3.extracted.mzML}
\alias{HAM005_641fE_14-11-07--Exp4.extracted.mzML}
\alias{HAM005_641fE_14-11-07--Exp5.extracted.mzML}
\title{FTICR-MS data files}
\description{
Direct injection  fourier-transform ion cyclotron resonance (FTICR) mass
spectrometry (MS) data files.

Files:
\itemize{
\item HAM004_641fE_14-11-07--Exp1.extracted.mzML
\item HAM004_641fE_14-11-07--Exp2.extracted.mzML
\item HAM004_641fE_14-11-07--Exp3.extracted.mzML
\item HAM004_641fE_14-11-07--Exp4.extracted.mzML
\item HAM004_641fE_14-11-07--Exp5.extracted.mzML
\item HAM005_641fE_14-11-07--Exp1.extracted.mzML
\item HAM005_641fE_14-11-07--Exp2.extracted.mzML
\item HAM005_641fE_14-11-07--Exp3.extracted.mzML
\item HAM005_641fE_14-11-07--Exp4.extracted.mzML
\item HAM005_641fE_14-11-07--Exp5.extracted.mzML
}
}
\author{
Mark Haid
}
