% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscDiaDda.R
\name{miscDdaDia}
\alias{miscDdaDia}
\alias{Christoforou_2016_TMT_DDA_FragPipe_Fraction1_psm.tsv}
\alias{Christoforou_2016_TMT_DDA_FragPipe_Fraction2_psm.tsv}
\alias{Christoforou_2016_TMT_DDA_MaxQuant_evidence.txt}
\alias{Christoforou_2016_TMT_DDA_sage_results.sage.tsv}
\alias{Christoforou_2016_TMT_DDA_sage_tmt.tsv}
\alias{Derks_2022_plex_DIA_DIANN_report_subset.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_FragPipe_A_1_psm.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_FragPipe_A_2_psm.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_FragPipe_A_3_psm.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_FragPipe_B_1_psm.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_FragPipe_B_2_psm.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_FragPipe_B_3_psm.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_MaxQuant_evidence.txt}
\alias{vanPuyvelde_2022_LFQ_DDA_MaxQuant_peptides.txt}
\alias{vanPuyvelde_2022_LFQ_DDA_MaxQuant_proteinGroups.txt}
\alias{vanPuyvelde_2022_LFQ_DDA_PEAKS_LFQ_report.csv}
\alias{vanPuyvelde_2022_LFQ_DDA_sage_lfq.tsv}
\alias{vanPuyvelde_2022_LFQ_DDA_sage_results.sage.tsv}
\alias{vanPuyvelde_2022_LFQ_DIA_DIANN_report.parquet}
\alias{vanPuyvelde_2022_LFQ_DIA_DIANN_report.tsv}
\title{Misc DDA and DIA datasets}
\description{
Various DDA and DIA experiments that were processed with different software
to illustrate the \code{QFeatures::readQFeatures()} function:
\itemize{
\item \verb{vanPuyvelde_2022_LFQ_*}: DDA and DIA data from van Puyvelde et
al. (2022). LFQ benchmarking experiment.
\item \verb{Christoforou_2016_TMT_DDA_*}: data from Christoforou et
al. (2016). hyperLOPIT spatial proteomics experiment.
\item \code{Derks_2022_plex_DIA_DIANN_report_subset.tsv}: subset from Derks et
al. (2022) plexDIA data.
}

Source files are available on Zenodo: https://zenodo.org/records/19137577
}
\references{
\itemize{
\item Van Puyvelde, B., Daled, S., Willems, S. et al. A comprehensive LFQ
benchmark dataset on modern day acquisition strategies in proteomics. Sci
Data 9, 126 (2022). https://doi.org/10.1038/s41597-022-01216-6.
\item Christoforou, A., Mulvey, C., Breckels, L. et al. A draft map of the mouse
pluripotent stem cell spatial proteome. Nat Commun 7, 9992
(2016). https://doi.org/10.1038/ncomms9992
\item Derks, J., Leduc, A., Wallmann, G. et al. Increasing the throughput of
sensitive proteomics by plexDIA. Nat Biotechnol 41, 50–59
(2023). https://doi.org/10.1038/s41587-022-01389-w
}
}
\author{
Kristina Gomoryova and Laurent Gatto
}
