% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRM.R
\name{MRM}
\alias{MRM}
\alias{MRM.standmix.5.mzML}
\title{Multiple Reaction Monitoring mode (MRM) test files}
\description{
\itemize{
\item MRM-standmix-5.mzML: sample from mouse brain acquired by HILIC
ESI-QqQ/MS in Dynamic multiple reaction monitoring mode (MRM). HPLC
system was a 1290 Infinity (Agilent Technologies) coupled to ion-Funnel
Triple quadrupole 6490 mass spectrometer (Agilent Technologies). This
file was contributed by Xavi Domingo-Almenara from the The Scripps
Research Institute, San Diego, CA.
}

Data location: \href{https://doi.org/10.5281/zenodo.18502866}{DOI:10.5281/zenodo.18502866}
}
\author{
Xavi Domingo-Almenara
}
