% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS3TMT.R
\name{MS3TMT}
\alias{MS3TMT}
\alias{MS3TMT10_01022016_32917.33481.mzML.gz}
\alias{MS3TMT10_01022016_32917-33481.mzML.gz}
\alias{MS3TMT11.mzML}
\alias{fdms3tmt11}
\alias{fdms3tmt11.rda}
\title{MS3 SPS TMT data}
\description{
\itemize{
\item \emph{MS3TMT10_01022016_32917-33481.mzML.gz}: A subset of 565 spectra from a
currenly unpublished TMT 10-plex experiment run on an Thermo Orbitrap
Lumos with synchronous precursor selection (SPS) MS3. Only the MS2 spectra
were centroided during convertion using \code{msconvert} (ProteoWizard release:
3.0.9283 (2016-1-11)) using vendor libraries.
\item \emph{MS3TMT11.mzML}: A subset of 994 spectra from a currenly unpublished
MS3 SPS TMT 11-plex experiment converted to mzML using
\code{msconvert}. The file contains 30, 482 and 482 MS1, MS2 and MS3
spectra, respectively. The MS1 spectra are in profile mode; other MS
levels are centroided. See 'Sensitive and Accurate Quantitation of
Phosphopeptides Using TMT Isobaric Labeling Technique' for details
about the acquisition method.
\item An feature data containing identification data is available with
\code{fdms3tmt11}, which can be used to directly update the feature data.
}
}
\author{
Laurent Gatto
}
