% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PXD000001.R
\name{PXD000001}
\alias{PXD000001}
\alias{TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid}
\alias{TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML.gz}
\alias{TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.20141210.mzid}
\alias{TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.20141210.mzML.gz}
\title{PXD000001 Proteomics Data}
\description{
The PXD000001 files are part of the first ProteomeXchange
submission (Vizcaíno J.A. et al, 2014), and contain the following
files.
\itemize{
\item TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML.gz:
an TMT6 6-plex LC-MSMS data containing 6 human spiked-in
proteins in a constant \emph{Erwinia carotovora} protein backgroud. The
data is described in more details in Gatto and Christoforou
(2013).
\item TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid:
generated searching the raw data against the \emph{Erwinia
carotovora} fasta database
}
}
\references{
\itemize{
\item Vizcaíno J.A. et al. \emph{ProteomeXchange: globally co-ordinated
proteomics data submission and dissemination}, Nature
Biotechnology 2014, 32,
223–226. http://www.ncbi.nlm.nih.gov/pubmed/24727771
\item Gatto L. and Christoforou A. \emph{Using R and Bioconductor for
proteomics data analysis}, Biochim Biophys Acta - Proteins and
Proteomics, 2013. http://www.ncbi.nlm.nih.gov/pubmed/23692960
}
}
\seealso{
The \href{https://bioconductor.org/packages/3.16/rpx}{rpx} package can
be used to access and download any PRIDE/ProteomeXchange files.
}
