% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derks2022plexDIA.R
\name{Report.Derks2022.plexDIA}
\alias{Report.Derks2022.plexDIA}
\alias{Report.Derks2022.plexDIA.tsv}
\title{Derks 2022 plexDIA data}
\description{
Single cell proteomics data acquired by the Slavov Lab using the
plexDIA protocol. It contains quantitative information from
pancreatic ductal acinar cells (PDAC; HPAF-II), melanoma cells
(WM989-A6-G3) and monocytes (U-937) at precursor and protein
level. The each run acquired 3 samples thanks to mTRAQ
multiplexing.

The data were downloaded from the Slavov lab google drive:
\itemize{
\item https://drive.google.com/drive/folders/1pUC2zgXKtKYn22mlor0lmUDK0frgwL_-
\item DIANN_outputs
\item wJD1146_1193_1200_tsvLib
\item Report.tsv
}

For more details about the data: https://plexdia.slavovlab.net/

The file is reshare here allow its dissemination via the MsDataHub
package.
}
\references{
Derks, J., Leduc, A., Wallmann, G. et al. Increasing the
throughput of sensitive proteomics by plexDIA. Nat Biotechnol
(2022). 10.1038/s41587-022-01389-w.
}
\author{
Laurent Gatto
}
