% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faahKO.R
\name{faahKO}
\alias{faahKO}
\alias{ko15.CDF}
\alias{ko16.CDF}
\alias{ko18.CDF}
\alias{ko19.CDF}
\alias{ko21.CDF}
\alias{ko22.CDF}
\alias{wt15.CDF}
\alias{wt16.CDF}
\alias{wt18.CDF}
\alias{wt19.CDF}
\alias{wt21.CDF}
\alias{wt22.CDF}
\title{Saghatelian et al. (2004) FAAH knock-out LC-MS data}
\description{
Data from the article
\href{http://dx.doi.org/10.1021/bi0480335}{Saghetalian et al.}: LC-MS data from 6
FAAH knock-out and 6 wild type mice acquired in positive ionization mode.
Data is subset to an m/z range of 200-600 and retention time range from
2500-4500 seconds. The MS data is provided as CDF files in (AIA/ANDI) NetCDF
format.

The data files are (6 knock-out, 6 wild-type samples):
\itemize{
\item \emph{ko15.CDF}
\item \emph{ko16.CDF}
\item \emph{ko18.CDF}
\item \emph{ko19.CDF}
\item \emph{ko21.CDF}
\item \emph{ko22.CDF}
\item \emph{wt15.CDF}
\item \emph{wt16.CDF}
\item \emph{wt18.CDF}
\item \emph{wt19.CDF}
\item \emph{wt21.CDF}
\item \emph{wt22.CDF}
}
}
\references{
\itemize{
\item Saghatelian, A et al. \emph{Assignment of endogenous substrates to enzymes
by global metabolite profiling}, Biochemistry, 2004.
http://dx.doi.org/10.1021/bi0480335
}
}
