% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsDataHub.R
\name{MsDataHub}
\alias{MsDataHub}
\title{All MsDataHub datasets}
\usage{
MsDataHub()
}
\value{
A \code{data.frame} describing the data available in
\code{MsDataHub}.
}
\description{
The \code{\link[=MsDataHub]{MsDataHub()}} function returns a \code{data.frame} with all the
annotated datasets provided in the package. For details on these
individual datasets, refer to their respective manual pages.
}
\examples{

MsDataHub()
}
\author{
Laurent Gatto
}
